Reorganize analysis endpoints
